requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12   
pydantic-settings==2.3.4
supabase==2.7.2
asyncpg==0.29.0
anthropic==0.34.2
google-generativeai==0.7.2
httpx==0.27.2
Pillow>=10.0.0
pytest==8.3.2
pytest-asyncio==0.24.0